pyproject.toml
[tool.poetry]
name = "rekognition_batch_detect"
version = "0.1.0"
description = ""
authors = ["Alessandro Cere <alecere@amazon.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
"aws-cdk.aws-s3" = "^1.108.1"
"aws-cdk.aws-sqs" = "^1.108.1"
"aws-cdk.aws-lambda" = "^1.108.1"
"aws-cdk.aws-lambda-event-sources" = "^1.108.1"
"aws-cdk.aws-lambda-python" = "^1.108.1"
"aws-cdk.aws-glue" = "^1.109.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
ipykernel = "^5.5.5"
tenacity = "^7.0.0"
aws-lambda-powertools = "^1.17.0"
requests = "^2.25.1"
s3fs = "^2021.6.0"
botostubs = "^0.14.1"
awswrangler = "^2.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
